wip entity api
